requirements.txt
flask
flask-cors
pymongo
python-dotenv
sentence-transformers
torch
